extract the functions outside
